+++ src/main/java/svenhjol/charmony/api/TotemType.java
@@ -2,6 +2,6 @@ package svenhjol.charmony.api;
 
 @SuppressWarnings("unused")
 public enum TotemType {
-    UNDYING,
-    PRESERVING
+    Undying,
+    Preserving
 }